Adding a horizontal border to the bottom of selected label items
Using the selectionIndicatorView doesn't seem to work for horizontal picker when used with auto layout, since the frame size of the picker is 0x0.  Rather than work through the details of supporting auto layout properly, we just added an optional bottom border width property which is applied to selected labels using the same color as the selected label text.